backend/requirements.txt
flask
flask_socketio
flask_cors
paho_mqtt
pytz